add xml.sax._exceptions module
related to https://github.com/python/typeshed/issues/3968
